Move GradesDetailsActivity to gradesdetails package
