+++ app/src/main/res/layout/activity_grades_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.grades.GradesDetailsActivity">
+    tools:context=".presentation.gradesdetails.GradesDetailsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/appBarLayoutGradesDetails"
